Move DayNavigator component to own file and move CSS styles with it
