+++ src/components/DayNavigator.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import './DayNavigator.css';
+
+const DayNavigator = ({dateDisplay, onPrev, onNext}) => {
+    return (
+        <nav className="page__nav">
+            <button
+                className="page__nav-button page__prev-day"
+                title="Go to previous day"
+                onClick={onPrev}
+            />
+            <h2 className="page__date">{dateDisplay}</h2>
+            <button
+                className="page__nav-button page__next-day"
+                title="Go to next day"
+                onClick={onNext}
+            />
+        </nav>
+    );
+};
+
+export default DayNavigator;
